Convert chart registry to a separate module
