+++ src/core/core.js
@@ -1,5 +1,5 @@
-import {constants} from './constants';
 import {config} from './config';
+import {chartRegistry} from './chart-registry';
 
 /**
  * The entire dc.js library is scoped under the **dc** name space. It does not introduce
@@ -18,165 +18,6 @@ import {config} from './config';
  *      .height(300)
  *      .filter('sunday');
  */
-
-/**
- * The dc.chartRegistry object maintains sets of all instantiated dc.js charts under named groups
- * and the default group.
- *
- * A chart group often corresponds to a crossfilter instance. It specifies
- * the set of charts which should be updated when a filter changes on one of the charts or when the
- * global functions {@link dc.filterAll dc.filterAll}, {@link dc.refocusAll dc.refocusAll},
- * {@link dc.renderAll dc.renderAll}, {@link dc.redrawAll dc.redrawAll}, or chart functions
- * {@link dc.baseMixin#renderGroup baseMixin.renderGroup},
- * {@link dc.baseMixin#redrawGroup baseMixin.redrawGroup} are called.
- *
- * @namespace chartRegistry
- * @memberof dc
- * @type {{has, register, deregister, clear, list}}
- */
-export const chartRegistry = (function () {
-    // chartGroup:string => charts:array
-    let _chartMap = {};
-
-    function initializeChartGroup (group) {
-        if (!group) {
-            group = constants.DEFAULT_CHART_GROUP;
-        }
-
-        if (!_chartMap[group]) {
-            _chartMap[group] = [];
-        }
-
-        return group;
-    }
-
-    return {
-        /**
-         * Determine if a given chart instance resides in any group in the registry.
-         * @method has
-         * @memberof dc.chartRegistry
-         * @param {Object} chart dc.js chart instance
-         * @returns {Boolean}
-         */
-        has: function (chart) {
-            for (let e in _chartMap) {
-                if (_chartMap[e].indexOf(chart) >= 0) {
-                    return true;
-                }
-            }
-            return false;
-        },
-
-        /**
-         * Add given chart instance to the given group, creating the group if necessary.
-         * If no group is provided, the default group `dc.constants.DEFAULT_CHART_GROUP` will be used.
-         * @method register
-         * @memberof dc.chartRegistry
-         * @param {Object} chart dc.js chart instance
-         * @param {String} [group] Group name
-         * @return {undefined}
-         */
-        register: function (chart, group) {
-            group = initializeChartGroup(group);
-            _chartMap[group].push(chart);
-        },
-
-        /**
-         * Remove given chart instance from the given group, creating the group if necessary.
-         * If no group is provided, the default group `dc.constants.DEFAULT_CHART_GROUP` will be used.
-         * @method deregister
-         * @memberof dc.chartRegistry
-         * @param {Object} chart dc.js chart instance
-         * @param {String} [group] Group name
-         * @return {undefined}
-         */
-        deregister: function (chart, group) {
-            group = initializeChartGroup(group);
-            for (let i = 0; i < _chartMap[group].length; i++) {
-                if (_chartMap[group][i].anchorName() === chart.anchorName()) {
-                    _chartMap[group].splice(i, 1);
-                    break;
-                }
-            }
-        },
-
-        /**
-         * Clear given group if one is provided, otherwise clears all groups.
-         * @method clear
-         * @memberof dc.chartRegistry
-         * @param {String} group Group name
-         * @return {undefined}
-         */
-        clear: function (group) {
-            if (group) {
-                delete _chartMap[group];
-            } else {
-                _chartMap = {};
-            }
-        },
-
-        /**
-         * Get an array of each chart instance in the given group.
-         * If no group is provided, the charts in the default group are returned.
-         * @method list
-         * @memberof dc.chartRegistry
-         * @param {String} [group] Group name
-         * @returns {Array<Object>}
-         */
-        list: function (group) {
-            group = initializeChartGroup(group);
-            return _chartMap[group];
-        }
-    };
-})();
-
-/**
- * Add given chart instance to the given group, creating the group if necessary.
- * If no group is provided, the default group `dc.constants.DEFAULT_CHART_GROUP` will be used.
- * @memberof dc
- * @method registerChart
- * @param {Object} chart dc.js chart instance
- * @param {String} [group] Group name
- * @return {undefined}
- */
-export const registerChart = function (chart, group) {
-    chartRegistry.register(chart, group);
-};
-
-/**
- * Remove given chart instance from the given group, creating the group if necessary.
- * If no group is provided, the default group `dc.constants.DEFAULT_CHART_GROUP` will be used.
- * @memberof dc
- * @method deregisterChart
- * @param {Object} chart dc.js chart instance
- * @param {String} [group] Group name
- * @return {undefined}
- */
-export const deregisterChart = function (chart, group) {
-    chartRegistry.deregister(chart, group);
-};
-
-/**
- * Determine if a given chart instance resides in any group in the registry.
- * @memberof dc
- * @method hasChart
- * @param {Object} chart dc.js chart instance
- * @returns {Boolean}
- */
-export const hasChart = function (chart) {
-    return chartRegistry.has(chart);
-};
-
-/**
- * Clear given group if one is provided, otherwise clears all groups.
- * @memberof dc
- * @method deregisterAllCharts
- * @param {String} group Group name
- * @return {undefined}
- */
-export const deregisterAllCharts = function (group) {
-    chartRegistry.clear(group);
-};
 
 /**
  * Clear all filters on all charts within the given chart group. If the chart group is not given then
